requirements.txt
public_ip
mysql-connector-python
colorama
matplotlib
pandas